Optimize site creation generated SQL Throwing TimeUtils functions

+++ src/main/java/org/craftercms/studio/impl/v2/utils/DependencyUtils.java
@@ -46,30 +46,55 @@ public class DependencyUtils {
     /**
      * Add the script snippets to update the dependencies for the given path
      *
-     * @param siteId  the site id
-     * @param path    the content item path
-     * @param oldPath the content item old path
-     * @param file    the file
+     * @param siteId            the site id
+     * @param path              the content item path
+     * @param oldPath           the content item old path
+     * @param file              the file
+     * @param dependencyService the dependency service
      * @throws IOException if an error occurs while updating the script
      */
-    public static void addDependenciesScriptSnippets(String siteId, String path, String oldPath, Path file, DependencyServiceInternal dependencyService)
+    public static void addDependenciesScriptSnippets(String siteId, String path, String oldPath,
+                                                     Path file, DependencyServiceInternal dependencyService)
             throws IOException, ServiceLayerException {
+        addDependenciesScriptSnippets(siteId, path, oldPath, file, dependencyService, true, true);
+    }
+
+    /**
+     * Add the script snippets to update the dependencies for the given path
+     *
+     * @param siteId            the site id
+     * @param path              the content item path
+     * @param oldPath           the content item old path
+     * @param file              the file
+     * @param dependencyService the dependency service
+     * @param cleanExisting     if true, the existing dependencies for the path will be deleted
+     * @param revalidate        if true, the existing dependencies pointing to the path will be set to valid=true
+     * @throws IOException if an error occurs while updating the script
+     */
+    public static void addDependenciesScriptSnippets(String siteId, String path, String oldPath,
+                                                     Path file, DependencyServiceInternal dependencyService,
+                                                     boolean cleanExisting, boolean revalidate)
+            throws IOException {
         Map<String, Set<ResolvedDependency>> dependencies = dependencyService.resolveDependencies(siteId, path);
-        if (isEmpty(oldPath)) {
-            Files.write(file, deleteDependencySourcePathRows(siteId, path).getBytes(UTF_8),
-                    StandardOpenOption.APPEND);
-            Files.write(file, "\n\n".getBytes(UTF_8), StandardOpenOption.APPEND);
-        } else {
-            Files.write(file, deleteDependencySourcePathRows(siteId, oldPath).getBytes(UTF_8),
-                    StandardOpenOption.APPEND);
-            // Invalidate existing dependencies pointing to the old item path
-            Files.write(file, invalidateDependencies(siteId, oldPath).getBytes(UTF_8),
-                    StandardOpenOption.APPEND);
+        if (cleanExisting) {
+            if (isEmpty(oldPath)) {
+                Files.write(file, deleteDependencySourcePathRows(siteId, path).getBytes(UTF_8),
+                        StandardOpenOption.APPEND);
+                Files.write(file, "\n\n".getBytes(UTF_8), StandardOpenOption.APPEND);
+            } else {
+                Files.write(file, deleteDependencySourcePathRows(siteId, oldPath).getBytes(UTF_8),
+                        StandardOpenOption.APPEND);
+                // Invalidate existing dependencies pointing to the old item path
+                Files.write(file, invalidateDependencies(siteId, oldPath).getBytes(UTF_8),
+                        StandardOpenOption.APPEND);
+            }
         }
 
-        // Validate existing broken dependencies pointing to the item path
-        Files.write(file, validateDependencies(siteId, path).getBytes(UTF_8),
-                StandardOpenOption.APPEND);
+        if (revalidate) {
+            // Validate existing broken dependencies pointing to the item path
+            Files.write(file, validateDependencies(siteId, path).getBytes(UTF_8),
+                    StandardOpenOption.APPEND);
+        }
 
         if (MapUtils.isEmpty(dependencies)) {
             return;
